Add Parser, Storage, TaskList, UI class and do basic categorisation
