+++ src/main/java/Parser.java
@@ -1,0 +1,93 @@
+public class Parser {
+
+    public static Task parseTask(String line) {
+        String[] parts = line.split(" \\| ");
+        if (parts.length >= 3) {
+            String type = parts[0].trim();
+            boolean isDone = parts[1].trim().equals("1");
+            String description = parts[2].trim();
+            switch (type) {
+                case "T":
+                    Task todo = new ToDo(description);
+                    if (isDone) todo.setIsDone();
+                    return todo;
+                case "D":
+                    String deadline = parts[3].trim();
+                    Task deadlineTask = new Deadline(description, deadline);
+                    if (isDone) deadlineTask.setIsDone();
+                    return deadlineTask;
+                case "E":
+                    String from = parts[3].trim();
+                    String to = parts[4].trim();
+                    Task eventTask = new Event(description, from, to);
+                    if (isDone) eventTask.setIsDone();
+                    return eventTask;
+                default:
+                    System.out.println("Unknown task type: " + type);
+            }
+        }
+        return null;
+    }
+
+    public static String taskToFileString(Task task) {
+        StringBuilder sb = new StringBuilder();
+        sb.append(task.type + " | ");
+        sb.append(task.isDone() ? "1 | " : "0 | ");
+        sb.append(task.description);
+        if (task instanceof Deadline) {
+            sb.append(" | ").append(((Deadline) task).by);
+        } else if (task instanceof Event) {
+            sb.append(" | ").append(((Event) task).from)
+                    .append(" | ").append(((Event) task).to);
+        }
+        return sb.toString();
+    }
+    public void processInput(String input, TaskList list, UI ui, Storage storage) throws XBotException {
+        String[] words = input.split("\\s+", 2);
+        String command = words[0].toLowerCase();
+        String rest = words.length > 1 ? words[1] : "";
+        switch(command) {
+            case "list":
+                ui.showTaskList(list);
+                break;
+            case "mark":
+                list.markDone(rest);
+                storage.saveTask(list);
+                break;
+            case "unmark":
+                list.markUndone(rest);
+                storage.saveTask(list);
+                break;
+            case "todo":
+                if (rest.isEmpty()) {
+                    throw new XBotException("The description of the todo cannot be empty!");
+                }
+                list.addTodo(rest);
+                storage.saveTask(list);
+                break;
+            case "event":
+                if (rest.isEmpty()) {
+                    throw new XBotException("The description of the event cannot be empty!");
+                }
+                list.addEvent(rest);
+                storage.saveTask(list);
+                break;
+            case "deadline":
+                if (rest.isEmpty()) {
+                    throw new XBotException("The description of the deadline cannot be empty!");
+                }
+                list.addDeadline(rest);
+                storage.saveTask(list);
+                break;
+            case "delete":
+                if (rest.isEmpty()) {
+                    throw new XBotException("The task number to be deleted cannot be empty!");
+                }
+                list.deleteTask(rest);
+                storage.saveTask(list);
+                break;
+            default:
+                throw new XBotException("I'm sorry, but I don't know what that means :-(");
+        }
+    }
+}
